% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odeintr.R
\name{set_optimization}
\alias{set_optimization}
\title{Set compiler optimization}
\usage{
set_optimization(level = 3, omit.debug = FALSE, overwrite = FALSE)
}
\arguments{
\item{level}{the compiler optimization level (-O<level>)}

\item{omit.debug}{if ture, remove "-g" from flags}

\item{overwrite}{if true, overwrite existing Makevars file}
}
\description{
Write a user Makevars with updated optimization level
}
\details{
This function will change the optimization flags used
when compiling code. It will write the file "Makevars" to the
".R" directory in your "$HOME" directory. These setting will
effect all subsequent compiles, even package installation,
unless you remove or edit the "Makevars" file.

This function assumes that your compiler uses "-O" to
indicate optimization level and "-g" to indicate that
the compiler should issue debugging symbols.
}
\note{
Don't go overboard. Levels greater than 3 can be
hazardous to numerical accuracy. Some packages will not
compile or will give inaccurate results for levels above 2.
}
\author{
Timothy H. Keitt
}
\seealso{
\code{\link{compile_sys}}
}

